add we are one group component
