+++ src/pages/components/ContactUsBanner/index.js
@@ -1,6 +1,6 @@
 import React from "react"
-import Box from '@material-ui/core/Box';
-import Grid from '@material-ui/core/Grid';
+import Box from '@material-ui/core/Box'
+import Grid from '@material-ui/core/Grid'
 
 const ContactUsBanner = () => {
   return (
