+++ src/pages/index.js
@@ -14,6 +14,7 @@ import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 
 import WhatWeDoSvg from "./SvgComponents/WhatWeDo";
 import ContactUsBanner from './components/ContactUsBanner'
+import WeAreOneGroup from './components/WeAreOneGroup'
 
 const Home = () => {
   const isMobile = useMediaQuery({ query: '(max-width: 767px)' })
@@ -496,65 +497,6 @@ const Home = () => {
     )
   }
 
-  const WeAreOneGroup = () => {
-    return (
-      <Box className={clsx("containerSec")}>
-        <Box className={clsx("sectionNoPadding")}>
-          <Box className={"h3Box"}>
-            <h3 style={{textAlign:'center'}}>We Are One Group</h3>
-          </Box>
-          <Box className={"spacingBox"}>
-            <Grid container>
-              <Grid item xs={12} md={3}>
-                <Box className={"imgBoxLogos"}>
-                  <a target="_blank" href="http://latamlink.io/">
-                    <img
-                      className={"oneGroupLogo"}
-                      alt="LatamLink"
-                      src={useBaseUrl("img/Group47.svg")}
-                    />
-                  </a>
-                </Box>
-              </Grid>
-              <Grid item xs={12} md={3}>
-                <Box className={"imgBoxLogos"}>
-                  <a target="_blank" href="https://www.edenia.com/">
-                    <img
-                      className={"oneGroupLogo"}
-                      alt="Edenia"
-                      src={useBaseUrl("img/Group17.svg")}
-                    />
-                  </a>
-                </Box>
-              </Grid>
-              <Grid item xs={12} md={3}>
-                <Box className={"imgBoxLogos"}>
-                  <a target="_blank" href="https://eosurf.com/">
-                    <img
-                      className={"oneGroupLogo"}
-                      alt="Eossurf"
-                      src={useBaseUrl("img/Group11.svg")}
-                    />
-                  </a>
-                </Box>
-              </Grid>
-              <Grid item xs={12} md={3}>
-                <Box className={"imgBoxLogos"}>
-                  <a target="_blank" href="https://ticoblockchain.cr/">
-                    <img
-                      className={"oneGroupLogo"}
-                      alt="Ticoblockchain"
-                      src={useBaseUrl("img/Group55.svg")}
-                    />
-                  </a>
-                </Box>
-              </Grid>
-            </Grid>
-          </Box>
-        </Box>
-      </Box>
-    )
-  }
 
   const SomeProjects = () => {
     return (
